Add intelligent timers switching
In the single timer mode, if the micro break is in 20 minutes, and the
rest brake is in 19 minutes, then the micro break timer is shown,
showing that one still has 20 minutes left before taking a break.

That's not the most useful information because one knows from the
beginning that one would have to take a break after 19 minutes.

This patch fixes this and similar cases by providing information about
the shortest time to take a break and the shortest time to end a break.